Refactor timestamps to use uint32_t

+++ components/philips_series_2200/philips_series_2200.h
@@ -128,8 +128,8 @@ namespace esphome
 #endif
 
         private:
-            long last_message_from_mainboard_time_ = 0;
-            long last_message_from_display_time_ = 0;
+            uint32_t last_message_from_mainboard_time_ = 0;
+            uint32_t last_message_from_display_time_ = 0;
 
             /// @brief reference to uart connected to the display unit
             uart::UARTDevice display_uart_;
